Introduce SubmitTransactionUiLogic for shared code SendFundsUiLogic/ErgoPaySigningUiLogic #79
