+++ android/src/main/java/org/ergoplatform/android/transactions/ErgoPaySigningViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.CoroutineScope
+import org.ergoplatform.transactions.TransactionResult
 import org.ergoplatform.uilogic.transactions.ErgoPaySigningUiLogic
 
 class ErgoPaySigningViewModel: ViewModel() {
@@ -16,6 +17,30 @@ class ErgoPaySigningViewModel: ViewModel() {
     inner class AndroidErgoPaySigningUiLogic: ErgoPaySigningUiLogic() {
         override val coroutineScope: CoroutineScope
             get() = viewModelScope
+
+        override fun notifyWalletStateLoaded() {
+            // TODO Ergo Pay show in UI
+        }
+
+        override fun notifyDerivedAddressChanged() {
+            // TODO Ergo Pay show in UI
+        }
+
+        override fun notifyUiLocked(locked: Boolean) {
+            TODO("Not yet implemented")
+        }
+
+        override fun notifyHasTxId(txId: String) {
+            TODO("Not yet implemented")
+        }
+
+        override fun notifyHasErgoTxResult(txResult: TransactionResult) {
+            TODO("Not yet implemented")
+        }
+
+        override fun notifyHasSigningPromptData(signingPrompt: String) {
+            TODO("Not yet implemented")
+        }
 
         override fun notifyStateChanged(newState: State) {
             _uiStateRefresh.postValue(newState)
